specs complete

+++ controllers/catController.js
@@ -2,56 +2,96 @@
 const db = require("../database/db");
 const Cat = db.models.cat;
 
-const Create = async ({ session, catName, picPath }) => {
+const Create = async ({ session, name, picPath }) => {
   if (!session || !session.userId) {
-    return {
+    throw new Error({
       status: 401,
       message: "You can not post Cat pics without logging in first!",
-    };
+    });
   }
-  console.log("session", session, "catName", catName, "picPath", picPath);
+
   const cat = await Cat.create({
-    name: catName,
+    name,
     media: picPath,
     userId: session.userId,
   });
+
   return cat;
 };
 
 const Get = async () => {
   const cats = await Cat.findAll();
+  return cats;
 };
 
-const GetUsersCats = async ({ id }) => {
-  const cats = await Cat.findAll({});
+const GetUsersCats = async ({ userId, session }) => {
+  if (session.userId !== userId) {
+    throw new Error("403");
+  }
+  const cats = Cat.findAll({
+    where: {
+      userId,
+    },
+  });
+  return cats;
 };
 
-const GetOne = async (id) => {
+const GetOne = async ({ id, session }) => {
   const cat = await Cat.findOne({
     where: {
       id,
     },
   });
-};
-
-const Update = async (id, picPath, payload) => {
-  const updatedCat = await Cat.update(
-    { media: picPath, ...body },
-    {
-      where: {
-        id: params.id,
-      },
+  if (cat) {
+    if (session.userId !== cat.userId) {
+      throw new Error("403");
+    } else {
+      return cat;
     }
-  );
+  } else {
+    throw new Error("404");
+  }
 };
 
-const Delete = async () => {
-  const deletedCat = await Cat.destroy({
+const Update = async ({ id, picPath, session, payload }) => {
+  const cat = await Cat.findOne({
     where: {
       id,
     },
   });
-  res.send(deletedCat.id);
+  if (cat) {
+    if (session.userId !== cat.userId) {
+      throw new Error("You are NOT authorized to update this cat.");
+    }
+    return await cat.update(
+      { media: picPath, ...payload },
+      {
+        where: {
+          id,
+        },
+      }
+    );
+  } else {
+    throw new Error("Cat does not exist!");
+  }
+};
+
+const Delete = async ({ id, session }) => {
+  const cat = await Cat.findOne({
+    where: {
+      id,
+    },
+  });
+  if (session.userId !== cat.userId) {
+    throw new Error("403");
+  }
+  await Cat.destroy({
+    where: {
+      id,
+    },
+  });
+
+  return cat;
 };
 
 module.exports = { Create, Get, GetUsersCats, GetOne, Update, Delete };
